apply feedback after call
